Version 1.0

+++ routes/budget.js
@@ -4,6 +4,7 @@ const Budget = require('../model/Budget');
 const Transaction = require('../model/Transaction');
 //Add for Private Routes
 const verify = require('./verifyToken');
+const {budgetValidation, transactionValidation, uuidValidation} = require('../validation');
 
 //Add verify for private methods
 //TODO:
@@ -45,6 +46,10 @@ router.get('/', verify, async (req,res) => {
 
 //Get specific Budget
 router.get('/:budgetId', verify, async (req,res) => {
+    //Validate UUID data
+    const { error } = uuidValidation(req.params);
+    if (error) return res.status(400).send(error.details[0].message);
+
     const ownerId = new mongoose.Types.ObjectId(req.user._id);
     const budgetId = new mongoose.Types.ObjectId(req.params.budgetId);
     budget = await Budget.aggregate([
@@ -80,6 +85,10 @@ router.get('/:budgetId', verify, async (req,res) => {
 
 //Add New Budget
 router.post('/', verify, async (req,res) => {
+    //Validate Budget data
+    const { error } = budgetValidation(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
+
     const budget = new Budget({
         owner: req.user._id,
         name: req.body.name,
@@ -97,6 +106,14 @@ router.post('/', verify, async (req,res) => {
 
 //Update specific budget
 router.put('/:budgetId', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+    
+    //Validate Budget data
+    const { error } = budgetValidation(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
+
     try {
         updatedBudget = await Budget.updateOne({ $and: [{_id: {$eq: req.params.budgetId}},{owner: {$eq: req.user._id}}]},
                                                 {$set: {name: req.body.name,
@@ -115,6 +132,10 @@ router.put('/:budgetId', verify, async (req,res) => {
 
 //Delete Specific Budget
 router.delete('/:budgetId', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
     deletedBudget = await Budget.deleteOne({ $and: [{_id: {$eq: req.params.budgetId}},{owner: {$eq: req.user._id}}]});
     if (deletedBudget && deletedBudget.deletedCount == 1) {
         deletedTrans = await Transaction.deleteMany({budget: req.params.budgetId});
@@ -129,6 +150,10 @@ router.delete('/:budgetId', verify, async (req,res) => {
 
 //Get specific Budget all Transactions
 router.get('/:budgetId/transactions', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
     try {
         budget = await Budget.findOne({_id: req.params.budgetId}).populate('transactions');
         res.send({'transactions': budget.transactions});
@@ -139,6 +164,14 @@ router.get('/:budgetId/transactions', verify, async (req,res) => {
 
 //Add New Transaction
 router.post('/:budgetId/transactions', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
+    //Validate Transaction data
+    const { error } = transactionValidation(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
+
     try {
         const transaction = new Transaction({
             owner: req.user._id,
@@ -158,6 +191,10 @@ router.post('/:budgetId/transactions', verify, async (req,res) => {
 
 //Get specific transaction
 router.get('/:budgetId/transactions/:transactionId', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
     try {
         const transaction = await Transaction.findOne({$and: [{_id: req.params.transactionId},
                                                             {budget: req.params.budgetId},
@@ -174,6 +211,14 @@ router.get('/:budgetId/transactions/:transactionId', verify, async (req,res) => 
 
 //Update specific transaction
 router.put('/:budgetId/transactions/:transactionId', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
+    //Validate Transaction data
+    const { error } = transactionValidation(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
+
     try {
         updatedTransaction = await Transaction.updateOne({ $and: [{_id: req.params.transactionId},{owner: req.user._id},{budget: req.params.budgetId}]},
                                                 {$set: {name: req.body.name,
@@ -190,6 +235,10 @@ router.put('/:budgetId/transactions/:transactionId', verify, async (req,res) => 
 
 //Delete Specific Transaction
 router.delete('/:budgetId/transactions/:transactionId', verify, async (req,res) => {
+    //Validate UUID data
+    const { uuidError } = uuidValidation(req.params);
+    if (uuidError) return res.status(400).send(error.details[0].message);
+
     try {
         deletedTransaction = await Transaction.deleteOne({ $and: [{_id: req.params.transactionId},{owner: req.user._id},{budget: req.params.budgetId}]});
         if (deletedTransaction && deletedTransaction.deletedCount == 1) {
